requirements.txt
maturin
basedpyright
black
pillow
